feat(failed): view table for aggregating owners
I was trying to create a view table using this reference
https://ponder.sh/docs/schema/views#usage. But apparently it's not
working and keep returning this error

BuildError: Schema validation failed:
view 'hourly_bucket.hour' is a non-column selected field.

Despite I am using the exact code with the example. So after trying to
downgrade the version and do some other approach, apparently I couldn't
fix it. So this commit is to demonstrate how to achieve that error code.

In the next commit, I will reverse and not use view table.