feat: add token metadata fetched from token uri to db
currently its using json column type. the problem with json column type
is that its not indexed and searchable. which is good for logging but
not good for query. in the next commit, I would like to use jsonb and
compare the write performance. and see if the indexing and search is
working. the other option is to add new columns and extract the value
from the json metadata. and just store the rest of metadata using json
as logging.